+++ packages/libs/react-components/src/typings/index.ts
@@ -1,34 +1,6 @@
-import { breakpoints, sizes } from '../styles/stitches.config';
-import { IThemeColors } from '..';
-
-import { ReactNode } from 'react';
+import { IThemeBreakpointProps } from '../styles';
 
 export type Primitive = number | string | boolean;
-export type IThemeSizes = `$${keyof typeof sizes}`;
-
-export type IThemeBreakpointProps = keyof typeof breakpoints;
-export type IStyleBreakpointProps = `@${keyof typeof breakpoints}`;
-
 export type ICSSPropType =
   | Primitive
   | Partial<Record<IThemeBreakpointProps, Primitive>>;
-
-export interface IGridContainerProps {
-  children?: ReactNode;
-  gap?: ICSSPropType;
-  templateAreas?: ICSSPropType;
-  templateRows?: ICSSPropType;
-  templateColumns?: ICSSPropType;
-}
-
-export type IColSpan = number | 'auto';
-
-export interface IGridItemProps {
-  children?: ReactNode;
-  colStart?: IColSpan;
-  colEnd?: IColSpan;
-  rowStart?: IColSpan;
-  rowEnd?: IColSpan;
-  area?: string;
-  bg?: IThemeColors;
-}
